added auto join
