+++ chaincode/resources/collections-config.json
@@ -1,0 +1,16 @@
+[
+    {
+        "name": "ibmResourcesPrivateData",
+        "policy": "OR('ibm.member')",
+        "requiredPeerCount": 1,
+        "maxPeerCount": 3,
+        "blockToLive": 0
+    },
+    {
+        "name": "oracleResourcesPrivateData",
+        "policy": "OR('oracle.member')",
+        "requiredPeerCount": 1,
+        "maxPeerCount": 3,
+        "blockToLive": 0
+    }
+]
